enh: Jacobian based penalty gradient computation [Transformation]
- Add pure virtual Penalty and DerivativeWrtJacobianDet to base class.
- Move code to evaluate penalty and compute gradient to base class.
- TopologyPreservation gradient computation using base class.
- Remove obsolete MinJacobianConstraint, use TopologyPreservation instead.
