+++ Modules/Common/include/mirtk/EnergyMeasure.h
@@ -1,8 +1,8 @@
 /*
  * Medical Image Registration ToolKit (MIRTK)
  *
- * Copyright 2013-2016 Imperial College London
- * Copyright 2013-2016 Andreas Schuh
+ * Copyright 2013-2017 Imperial College London
+ * Copyright 2013-2017 Andreas Schuh
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -103,7 +103,6 @@ enum EnergyMeasure
     EM_L1Norm,                  ///< Sparsity constraint based on l1-norm
     EM_L2Norm,                  ///< Sparsity constraint based on l2-norm
     EM_SqLogDetJac,             ///< Squared logarithm of the Jacobian determinant
-    EM_MinDetJac,               ///< Constrain minimum Jacobian determinant
 
   CM_End,
 
@@ -179,7 +178,6 @@ inline string ToString(const EnergyMeasure &value, int w, char c, bool left)
     case EM_L1Norm:               str = "L1"; break;
     case EM_L2Norm:               str = "L2"; break;
     case EM_SqLogDetJac:          str = "SqLogDetJac"; break;
-    case EM_MinDetJac:            str = "MinDetJac"; break;
 
     // -------------------------------------------------------------------------
     // Others
@@ -254,7 +252,6 @@ inline string ToPrettyString(const EnergyMeasure &value, int w = 0, char c = ' '
     case EM_L1Norm:               str = "l1 norm"; break;
     case EM_L2Norm:               str = "l2 norm"; break;
     case EM_SqLogDetJac:          str = "Squared logarithm of Jacobian determinant"; break;
-    case EM_MinDetJac:            str = "Minimum Jacobian determinant"; break;
 
     // -------------------------------------------------------------------------
     // Others
@@ -330,8 +327,7 @@ inline bool FromString(const char *str, EnergyMeasure &value)
   // ---------------------------------------------------------------------------
   // Alternative names for transformation regularization terms
   if (value == EM_Unknown) {
-    if      (lstr == "jac")    value = EM_SqLogDetJac;
-    else if (lstr == "minjac") value = EM_MinDetJac;
+    if (lstr == "jac") value = EM_SqLogDetJac;
   }
 
   // ---------------------------------------------------------------------------
