+++ Modules/Transformation/src/CMakeLists.txt
@@ -46,7 +46,6 @@ set(HEADERS
   LinearFreeFormTransformation4D.h
   LinearFreeFormTransformationTD.h
   LogJacobianConstraint.h
-  MinJacobianConstraint.h
   MultiLevelFreeFormTransformation.h
   MultiLevelStationaryVelocityTransformation.h
   MultiLevelTransformation.h
@@ -93,7 +92,6 @@ set(SOURCES
   LinearFreeFormTransformation4D.cc
   LinearFreeFormTransformationTD.cc
   LogJacobianConstraint.cc
-  MinJacobianConstraint.cc
   MultiLevelFreeFormTransformation.cc
   MultiLevelStationaryVelocityTransformation.cc
   MultiLevelTransformation.cc
