+++ Modules/Transformation/src/TopologyPreservationConstraint.cc
@@ -1,8 +1,8 @@
 /*
  * Medical Image Registration ToolKit (MIRTK)
  *
- * Copyright 2013-2015 Imperial College London
- * Copyright 2013-2015 Andreas Schuh
+ * Copyright 2013-2017 Imperial College London
+ * Copyright 2013-2017 Andreas Schuh
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,9 +19,6 @@
 
 #include "mirtk/TopologyPreservationConstraint.h"
 
-#include "mirtk/FreeFormTransformation.h"
-#include "mirtk/ObjectFactory.h"
-
 
 namespace mirtk {
 
@@ -33,50 +30,33 @@ mirtkAutoRegisterEnergyTermMacro(TopologyPreservationConstraint);
 // -----------------------------------------------------------------------------
 TopologyPreservationConstraint::TopologyPreservationConstraint(const char *name)
 :
-  TransformationConstraint(name)
+  JacobianConstraint(name),
+  _Gamma(.3)
 {
   _ConstrainPassiveDoFs = true;
 }
 
+
+// =============================================================================
+// Parameters
+// =============================================================================
+
 // -----------------------------------------------------------------------------
-double TopologyPreservationConstraint::Evaluate()
+bool TopologyPreservationConstraint::SetWithoutPrefix(const char *param, const char *value)
 {
-  const FreeFormTransformation *ffd = FFD();
-  if (!ffd) return .0; // No penalty for non-deformable transformation
-
-  double penalty = .0;
-  int    nactive = 0;
-
-  double x, y, z, det;
-  for (int l = -1; l < ffd->T(); ++l)
-  for (int k = -1; k < ffd->Z(); ++k)
-  for (int j = -1; j < ffd->Y(); ++j)
-  for (int i = -1; i < ffd->X(); ++i) {
-    if (_ConstrainPassiveDoFs || (i >= 0 && j >= 0 && k >= 0 && l >= 0 && ffd->IsActive(i, j, k, l))) {
-      x = i + 0.5;
-      y = j + 0.5;
-      z = k + 0.5;
-      ffd->LatticeToWorld(x, y, z);
-      det = ffd->LocalJacobian(x, y, z, ffd->LatticeToTime(l + 0.5));
-      if (det < 0.3) {
-        // 10 * det(J)^2 + 1 / (10 * det(J)^2) - 2
-        det     *= 10.0 * det;
-        penalty += det + 1.0 / det - 2.0;
-      }
-      ++nactive;
-    }
+  if (strcmp(param, "Threshold") == 0 ||
+      strcmp(param, "Gamma")     == 0) {
+    return FromString(value, _Gamma);
   }
-
-  if (nactive) penalty /= nactive;
-  return penalty;
+  return JacobianConstraint::SetWithoutPrefix(param, value);
 }
 
 // -----------------------------------------------------------------------------
-void TopologyPreservationConstraint::EvaluateGradient(double *, double, double)
+ParameterList TopologyPreservationConstraint::Parameter() const
 {
-  // TODO: Implement gradient computation (c.f. VolumePreservationConstraint::EvaluateGradient)
-  cerr << "TopologyPreservationConstraint::EvaluateGradient: Not implemented yet" << endl;
-  exit(1);
+  ParameterList params = JacobianConstraint::Parameter();
+  InsertWithPrefix(params, "Threshold", _Gamma);
+  return params;
 }
